[DNM] Retrigger CI build
Looking at the 1.3.0 release build the build failed during the audit
stage checking the shared object file that was built. This was failing
because the macOS target version in the binary was 10.12 but the
environment variable to override the default target of 10.9 was not
getting set properly. We set this in the pyproject.toml but it is
getting overridden by the PGO settings via CIBW_ENVIRONMENT in the
github actions config. If you look at the recent builds we were
publishing 10.9 target wheels on pypi despite documenting we target
10.12. The CI job logs also show that env variable is never being
set during the cibuildwheel process.

The release of rust 1.83 in the last 24 hrs probably is what changed
the output here causing the failure, although rust was supposed to
raise it's minimum target to 10.12 in 1.74 [1] something probably
changed in the release today which is tripping up the audit now.

This PR is temporary to retrigger CI with this fix, we can open a real
PR fixing this will be published after the 1.3.0 release is live.

[1] https://blog.rust-lang.org/2023/09/25/Increasing-Apple-Version-Requirements.html